soc: ambiq: apollo3x: Initial support for Apollo3 Blue Plus
Added using the Ambiq apollo3 HAL along with some basic changes
for the different hardware in the GPIO, STIMER, and PINCTRL
drivers
