+++ boards/arm/apollo3p_blue_evb/apollo3p_blue_evb-pinctrl.dtsi
@@ -9,31 +9,170 @@
 #include "apollo3p_blue_evb_connector.dtsi"
 
 &pinctrl {
-	uart0_default: uart0_default { // done
+	uart0_default: uart0_default {
 		group1 {
-			pinmux = <UART0TX_P23>;
+			pinmux = <UART0TX_P22>;
 		};
 		group2 {
-			pinmux = <UART0RX_P22>;
+			pinmux = <UART0RX_P23>;
 			input-enable;
 		};
 	};
-	i2c0_default: i2c0_default { // done
+	uart1_default: uart1_default {
+		group1 {
+			pinmux = <UART1TX_P35>;
+		};
+		group2 {
+			pinmux = <UART1RX_P36>;
+			input-enable;
+		};
+	};
+
+	i2c0_default: i2c0_default {
+		group1 {
+			pinmux = <M0SCL_P5>, <M0SDAWIR3_P6>;
+			drive-open-drain;
+			drive-strength = <2>;
+			bias-pull-up;
+		};
+	};
+	i2c1_default: i2c1_default {
 		group1 {
 			pinmux = <M1SCL_P8>, <M1SDAWIR3_P9>;
 			drive-open-drain;
-			drive-strength = "0.5";
+			drive-strength = <2>;
+			bias-pull-up;
+		};
+	};
+	i2c2_default: i2c2_default {
+		group1 {
+			pinmux = <M2SCL_P27>, <M2SDAWIR3_P25>;
+			drive-open-drain;
+			drive-strength = <2>;
+			bias-pull-up;
+		};
+	};
+	i2c3_default: i2c3_default {
+		group1 {
+			pinmux = <M3SCL_P42>, <M3SDAWIR3_P43>;
+			drive-open-drain;
+			drive-strength = <2>;
+			bias-pull-up;
+		};
+	};
+	i2c4_default: i2c4_default {
+		group1 {
+			pinmux = <M4SCL_P39>, <M4SDAWIR3_P40>;
+			drive-open-drain;
+			drive-strength = <2>;
+			bias-pull-up;
+		};
+	};
+	i2c5_default: i2c5_default {
+		group1 {
+			pinmux = <M5SCL_P48>, <M5SDAWIR3_P49>;
+			drive-open-drain;
+			drive-strength = <2>;
 			bias-pull-up;
 		};
 	};
 
-	spi0_default: spi0_default { // done
+	spi0_default: spi0_default {
 		group1 {
-			pinmux = <M0SCK_P5>, <M0MISO_P7>, <M0MOSI_P6>;
+			pinmux = <M0SCK_P5>, <M0MISO_P6>, <M0MOSI_P7>;
 		};
 		group2 {
-			pinmux = <NCE42_P42>;
+			pinmux = <NCE11_P11>;
 			drive-push-pull;
+			ambiq,iom-nce-module = <0>;
+		};
+	};
+	spi1_default: spi1_default {
+		group1 {
+			pinmux = <M1SCK_P8>, <M1MISO_P9>, <M1MOSI_P10>;
+		};
+		group2 {
+			pinmux = <NCE34_P34>;
+			drive-push-pull;
+			ambiq,iom-nce-module = <2>;
+		};
+	};
+	spi2_default: spi2_default {
+		group1 {
+			pinmux = <M2SCK_P27>, <M2MISO_P25>, <M2MOSI_P28>;
+		};
+		group2 {
+			pinmux = <NCE15_P15>;
+			drive-push-pull;
+			ambiq,iom-nce-module = <3>;
+		};
+	};
+	spi3_default: spi3_default {
+		group1 {
+			pinmux = <M3SCK_P42>, <M3MISO_P43>, <M3MOSI_P38>;
+		};
+		group2 {
+			pinmux = <NCE29_P29>;
+			drive-push-pull;
+			ambiq,iom-nce-module = <0>;
+		};
+	};
+	spi4_default: spi4_default {
+		group1 {
+			pinmux = <M4SCK_P39>, <M4MISO_P40>, <M4MOSI_P44>;
+		};
+		group2 {
+			pinmux = <NCE13_P13>;
+			drive-push-pull;
+			ambiq,iom-nce-module = <1>;
+		};
+	};
+	spi5_default: spi5_default {
+		group1 {
+			pinmux = <M5SCK_P48>, <M5MISO_P49>, <M5MOSI_P47>;
+		};
+		group2 {
+			pinmux = <NCE16_P16>;
+			drive-push-pull;
+			ambiq,iom-nce-module = <0>;
+		};
+	};
+	mspi0_default: mspi0_default{
+		group1 {
+			pinmux = <MSPI0_0_P22>,
+				 <MSPI0_1_P26>,
+				 <MSPI0_8_P24>;
+		};
+		group2 {
+			pinmux = <NCE28_P28>;
+			drive-push-pull;
+			drive-strength = <2>;
+			ambiq,iom-nce-module = <0>;
+		};
+	};
+	mspi1_default: mspi1_default{
+		group1 {
+			pinmux = <MSPI1_0_P51>,
+				 <MSPI1_1_P52>,
+				 <MSPI1_8_P59>;
+		};
+		group2 {
+			pinmux = <NCE50_P50>;
+			drive-push-pull;
+			drive-strength = <2>;
+			ambiq,iom-nce-module = <0>;
+		};
+	};
+	mspi2_default: mspi2_default{
+		group1 {
+			pinmux = <MSPI2_0_P64>,
+				 <MSPI2_1_P65>,
+				 <MSPI2_4_P68>;
+		};
+		group2 {
+			pinmux = <NCE63_P63>;
+			drive-push-pull;
+			drive-strength = <2>;
 			ambiq,iom-nce-module = <0>;
 		};
 	};
