+++ boards/arm/apollo3p_blue_evb/apollo3p_blue_evb.dts
@@ -1,5 +1,5 @@
 /dts-v1/;
-#include <ambiq/ambiq_apollo3p.dtsi>
+#include <ambiq/ambiq_apollo3p_blue.dtsi>
 
 #include "apollo3p_blue_evb-pinctrl.dtsi"
 
@@ -29,11 +29,11 @@
 	leds {
 		compatible = "gpio-leds";
 		led0: led_0 {
-			gpios = <&gpio0_31 17 GPIO_ACTIVE_LOW>;
+			gpios = <&gpio0_31 10 GPIO_ACTIVE_LOW>;
 			label = "LED 0";
 		};
 		led1: led_1 {
-			gpios = <&gpio0_31 14 GPIO_ACTIVE_LOW>;
+			gpios = <&gpio0_31 30 GPIO_ACTIVE_LOW>;
 			label = "LED 1";
 		};
 		led2: led_2 {
@@ -41,11 +41,11 @@
 			label = "LED 2";
 		};
 		led3: led_3 {
-			gpios = <&gpio0_31 30 GPIO_ACTIVE_LOW>;
+			gpios = <&gpio0_31 14 GPIO_ACTIVE_LOW>;
 			label = "LED 3";
 		};
 		led4: led_4 {
-			gpios = <&gpio0_31 10 GPIO_ACTIVE_LOW>;
+			gpios = <&gpio0_31 17 GPIO_ACTIVE_LOW>;
 			label = "LED 4";
 		};
 	};
