+++ drivers/gpio/CMakeLists.txt
@@ -90,6 +90,8 @@ zephyr_library_sources_ifdef(CONFIG_GPIO_BCM2711    gpio_bcm2711.c)
 zephyr_library_sources_ifdef(CONFIG_GPIO_RA         gpio_ra.c)
 zephyr_library_sources_ifdef(CONFIG_GPIO_RZT2M    gpio_rzt2m.c)
 zephyr_library_sources_ifdef(CONFIG_GPIO_AMBIQ      gpio_ambiq.c)
+zephyr_library_sources_ifdef(CONFIG_SOC_SERIES_APOLLO4X    gpio_ambiq_apollo4x.c)
+zephyr_library_sources_ifdef(CONFIG_SOC_SERIES_APOLLO3X    gpio_ambiq_apollo3x.c)
 
 if (CONFIG_GPIO_EMUL_SDL)
   zephyr_library_sources(gpio_emul_sdl.c)
