+++ drivers/pinctrl/CMakeLists.txt
@@ -5,7 +5,7 @@ zephyr_library()
 zephyr_library_sources(common.c)
 zephyr_library_sources_ifdef(CONFIG_PINCTRL_TELINK_B91 pinctrl_b91.c)
 zephyr_library_sources_ifdef(CONFIG_PINCTRL_AMBIQ_APOLLO4 pinctrl_ambiq_apollo4.c)
-zephyr_library_sources_ifdef(CONFIG_PINCTRL_AMBIQ_APOLLO3 pinctrl_ambiq_apollo4.c)
+zephyr_library_sources_ifdef(CONFIG_PINCTRL_AMBIQ_APOLLO3 pinctrl_ambiq_apollo3.c)
 zephyr_library_sources_ifdef(CONFIG_PINCTRL_GD32_AF pinctrl_gd32_af.c)
 zephyr_library_sources_ifdef(CONFIG_PINCTRL_GD32_AFIO pinctrl_gd32_afio.c)
 zephyr_library_sources_ifdef(CONFIG_PINCTRL_ITE_IT8XXX2 pinctrl_ite_it8xxx2.c)
