+++ include/zephyr/dt-bindings/pinctrl/ambiq-apollo3-pinctrl.h
@@ -35,7 +35,7 @@
 #define MSPI0_6_P2 APOLLO3_PINMUX(2, 5)
 #define NCE2_P2 APOLLO3_PINMUX(2, 7)
 #define UA0RTS_P3 APOLLO3_PINMUX(3, 0)
-#define SLnCE_P3 APOLLO3_PINMUX(3, 1)
+#define SLNCE_P3 APOLLO3_PINMUX(3, 1)
 #define NCE3_P3 APOLLO3_PINMUX(3, 2)
 #define GPIO3_P3 APOLLO3_PINMUX(3, 3)
 #define MSPI0_7_P3 APOLLO3_PINMUX(3, 5)
@@ -48,19 +48,19 @@
 #define UART1RX_P4 APOLLO3_PINMUX(4, 5)
 #define CT17_P4 APOLLO3_PINMUX(4, 6)
 #define MSPI0_2_P4 APOLLO3_PINMUX(4, 7)
-#define MOSCL_P5 APOLLO3_PINMUX(5, 0)
-#define MOSCK_P5 APOLLO3_PINMUX(5, 1)
+#define M0SCL_P5 APOLLO3_PINMUX(5, 0)
+#define M0SCK_P5 APOLLO3_PINMUX(5, 1)
 #define UA0RTS_P5 APOLLO3_PINMUX(5, 2)
 #define GPIO5_P5 APOLLO3_PINMUX(5, 3)
 #define CT8_P5 APOLLO3_PINMUX(5, 7)
-#define MOSDAWIR3_P6 APOLLO3_PINMUX(6, 0)
-#define MOMISO_P6 APOLLO3_PINMUX(6, 1)
+#define M0SDAWIR3_P6 APOLLO3_PINMUX(6, 0)
+#define M0MISO_P6 APOLLO3_PINMUX(6, 1)
 #define UA0CTS_P6 APOLLO3_PINMUX(6, 2)
 #define GPIO6_P6 APOLLO3_PINMUX(6, 3)
 #define CT10_P6 APOLLO3_PINMUX(6, 5)
 #define I2SDAT_P6 APOLLO3_PINMUX(6, 7)
 #define NCE7_P7 APOLLO3_PINMUX(7, 0)
-#define MOMOSI_P7 APOLLO3_PINMUX(7, 1)
+#define M0MOSI_P7 APOLLO3_PINMUX(7, 1)
 #define CLKOUT_P7 APOLLO3_PINMUX(7, 2)
 #define GPIO7_P7 APOLLO3_PINMUX(7, 3)
 #define TRIG0_P7 APOLLO3_PINMUX(7, 4)
@@ -112,7 +112,7 @@
 #define GPIO14_P14 APOLLO3_PINMUX(14, 3)
 #define PDMCLK_P14 APOLLO3_PINMUX(14, 4)
 #define SWDCK_P14 APOLLO3_PINMUX(14, 6)
-#define 32KHzXT_P14 APOLLO3_PINMUX(14, 7)
+#define XT32KHZ_P14 APOLLO3_PINMUX(14, 7)
 #define ADCD1N_P15 APOLLO3_PINMUX(15, 0)
 #define NCE15_P15 APOLLO3_PINMUX(15, 1)
 #define UART1RX_P15 APOLLO3_PINMUX(15, 2)
@@ -183,7 +183,7 @@
 #define GPIO24_P24 APOLLO3_PINMUX(24, 3)
 #define UA0CTS_P24 APOLLO3_PINMUX(24, 4)
 #define CT21_P24 APOLLO3_PINMUX(24, 5)
-#define 32KHzXT_P24 APOLLO3_PINMUX(24, 6)
+#define XT32KHZ_P24 APOLLO3_PINMUX(24, 6)
 #define SWO_P24 APOLLO3_PINMUX(24, 7)
 #define UART1RX_P25 APOLLO3_PINMUX(25, 0)
 #define NCE25_P25 APOLLO3_PINMUX(25, 1)
@@ -242,7 +242,7 @@
 #define UA1CTS_P32 APOLLO3_PINMUX(32, 7)
 #define ADCSE5_P33 APOLLO3_PINMUX(33, 0)
 #define NCE33_P33 APOLLO3_PINMUX(33, 1)
-#define 32KHzXT_P33 APOLLO3_PINMUX(33, 2)
+#define XT32KHZ_P33 APOLLO3_PINMUX(33, 2)
 #define GPIO33_P33 APOLLO3_PINMUX(33, 3)
 #define BLEIF_CSN_P33 APOLLO3_PINMUX(33, 4)
 #define UA0CTS_P33 APOLLO3_PINMUX(33, 5)
@@ -268,7 +268,7 @@
 #define NCE36_P36 APOLLO3_PINMUX(36, 1)
 #define UART1RX_P36 APOLLO3_PINMUX(36, 2)
 #define GPIO36_P36 APOLLO3_PINMUX(36, 3)
-#define 32KHzXT_P36 APOLLO3_PINMUX(36, 4)
+#define XT32KHZ_P36 APOLLO3_PINMUX(36, 4)
 #define UA1CTS_P36 APOLLO3_PINMUX(36, 5)
 #define UA0CTS_P36 APOLLO3_PINMUX(36, 6)
 #define PDMDATA_P36 APOLLO3_PINMUX(36, 7)
@@ -340,7 +340,7 @@
 #define PDMCLK_P46 APOLLO3_PINMUX(46, 5)
 #define UART1TX_P46 APOLLO3_PINMUX(46, 6)
 #define SWO_P46 APOLLO3_PINMUX(46, 7)
-#define 32KHzXT_P47 APOLLO3_PINMUX(47, 0)
+#define XT32KHZ_P47 APOLLO3_PINMUX(47, 0)
 #define NCE47_P47 APOLLO3_PINMUX(47, 1)
 #define CT26_P47 APOLLO3_PINMUX(47, 2)
 #define GPIO47_P47 APOLLO3_PINMUX(47, 3)
