+++ soc/arm/ambiq/apollo3x/Kconfig.defconfig.apollo3p_blue
@@ -5,7 +5,7 @@
 if SOC_APOLLO3P_BLUE
 
 config NUM_IRQS
-	default 31
+	default 33 
 
 DT_NODE_SRAM := /memory@0
